+++ src/main/kotlin/no/nav/syfo/common/token/OboTokenProvider.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.common.azure
+package no.nav.syfo.common.token
 
 fun interface OboTokenProvider {
     suspend fun getOnBehalfOfToken(scopeClientId: String, token: String): String?