+++ src/test/kotlin/no/nav/syfo/common/tilgangskontroll/client/VeilederPersonerAccessTest.kt
@@ -8,7 +8,7 @@ import io.mockk.clearAllMocks
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
-import no.nav.syfo.common.azure.OboTokenProvider
+import no.nav.syfo.common.token.OboTokenProvider
 import no.nav.syfo.common.http.commonConfig
 import no.nav.syfo.common.testhelper.respond
 import org.junit.jupiter.api.AfterEach